Fixed major bug with inheritance.
The only check for whether or not to re-load the exemplars was in .spawn, but .gather_exemplars would go up the inheritance chain to get generators from parent classes. This caused some problems if a parent class had already gathered exemplars and then a child class caused it to re-gather them. (It was fine if the child class happened first.)

Moved the check to .gather_exemplars, added/moved specs, un-protected .gather_exemplars.